Guardando cambios

+++ app/model/UsuarioAdo.php
@@ -16,7 +16,7 @@ class UsuarioAdo
             $arrayUsuarios = array();
             $comandoUsuarios = Database::getInstance()->getDb()->prepare("SELECT 
             u.idUsuario, upper(u.Nombres) AS Nombres, upper(u.Apellidos) AS Apellidos, 
-            upper(u.Usuario) AS Usuario, u.Clave,r.Nombre as Rol FROM Usuario as u INNER JOIN Rol AS r ON r.idRol = u.Rol
+            upper(u.Usuario) AS Usuario, u.Clave,r.Nombre as Rol, u.Estado FROM Usuario as u INNER JOIN Rol AS r ON r.idRol = u.Rol
             where u.Nombres like concat('%', ?,'%') or u.Apellidos like concat('%', ?,'%') or u.Usuario like concat('%', ?,'%')
             ORDER BY u.idUsuario ASC
             offset ? rows fetch next ? rows only");
@@ -36,7 +36,8 @@ class UsuarioAdo
                     "Apellidos" => $row["Apellidos"],
                     "Usuario" => $row["Usuario"],
                     "Clave" => $row["Clave"],
-                    "Rol"=>$row["Rol"]
+                    "Rol" => $row["Rol"],
+                    "Estado" => $row["Estado"]
                 ));
             }
 
@@ -74,13 +75,14 @@ class UsuarioAdo
                     Database::getInstance()->getDb()->rollback();
                     return "duplicado";
                 } else {
-                    $comandoInsert = Database::getInstance()->getDb()->prepare("UPDATE Usuario SET Nombres = UPPER(?), Apellidos = UPPER(?), Usuario = UPPER(?), Clave = ?,Rol=? WHERE idUsuario = ?");
+                    $comandoInsert = Database::getInstance()->getDb()->prepare("UPDATE Usuario SET Nombres = UPPER(?), Apellidos = UPPER(?), Usuario = UPPER(?), Clave = ?,Rol = ?,Estado = ? WHERE idUsuario = ?");
                     $comandoInsert->bindParam(1, $usuario["nombres"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(2, $usuario["apellidos"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(3, $usuario["usuarios"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(4, $usuario["contrasena"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(5, $usuario["rol"], PDO::PARAM_INT);
-                    $comandoInsert->bindParam(6, $usuario["idusuario"], PDO::PARAM_INT);
+                    $comandoInsert->bindParam(6, $usuario["estado"], PDO::PARAM_INT);
+                    $comandoInsert->bindParam(7, $usuario["idusuario"], PDO::PARAM_INT);
                     $comandoInsert->execute();
                     Database::getInstance()->getDb()->commit();
                     return "actualizado";
@@ -95,12 +97,13 @@ class UsuarioAdo
                     Database::getInstance()->getDb()->rollback();
                     return "duplicado";
                 } else {
-                    $comandoInsert = Database::getInstance()->getDb()->prepare("INSERT INTO Usuario (Nombres, Apellidos, Usuario,Clave,Rol,Estado,Sistema) VALUES (UPPER(?), UPPER(?), UPPER(?), ?, ?, 1, 0)");
+                    $comandoInsert = Database::getInstance()->getDb()->prepare("INSERT INTO Usuario (Nombres, Apellidos, Usuario,Clave,Rol,Estado,Sistema) VALUES (UPPER(?), UPPER(?), UPPER(?), ?, ?, ?, 0)");
                     $comandoInsert->bindParam(1, $usuario["nombres"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(2, $usuario["apellidos"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(3, $usuario["usuarios"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(4, $usuario["contrasena"], PDO::PARAM_STR);
                     $comandoInsert->bindParam(5, $usuario["rol"], PDO::PARAM_INT);
+                    $comandoInsert->bindParam(6, $usuario["estado"], PDO::PARAM_INT);
                     $comandoInsert->execute();
                     Database::getInstance()->getDb()->commit();
                     return "insertado";
@@ -130,11 +133,20 @@ class UsuarioAdo
                     Database::getInstance()->getDb()->rollback();
                     return "activo";
                 } else {
-                    $comandoInsert = Database::getInstance()->getDb()->prepare("DELETE FROM Usuario WHERE idUsuario = ?");
-                    $comandoInsert->bindParam(1, $usuario["idusuario"], PDO::PARAM_STR);
-                    $comandoInsert->execute();
-                    Database::getInstance()->getDb()->commit();
-                    return "eliminado";
+
+                    $comandSelect = Database::getInstance()->getDb()->prepare("SELECT * FROM Ingreso WHERE idUsuario = ? ");
+                    $comandSelect->bindParam(1, $usuario["idusuario"], PDO::PARAM_STR);
+                    $comandSelect->execute();
+                    if ($comandSelect->fetch()) {
+                        Database::getInstance()->getDb()->rollback();
+                        return "ingreso";
+                    } else {
+                        $comandoInsert = Database::getInstance()->getDb()->prepare("DELETE FROM Usuario WHERE idUsuario = ?");
+                        $comandoInsert->bindParam(1, $usuario["idusuario"], PDO::PARAM_STR);
+                        $comandoInsert->execute();
+                        Database::getInstance()->getDb()->commit();
+                        return "eliminado";
+                    }
                 }
             }
         } catch (Exception $ex) {
@@ -162,7 +174,7 @@ class UsuarioAdo
             $array = array();
             $cmdLogin = Database::getInstance()->getDb()->prepare("SELECT u.idUsuario,u.Nombres,u.Apellidos,u.Usuario,u.Rol,r.Nombre,u.Estado,u.Sistema FROM Usuario as u  inner join Rol as r
             on u.Rol = r.idRol 
-            WHERE u.Usuario = ? AND u.Clave = ?");
+            WHERE u.Usuario = ? AND u.Clave = ? AND Estado = 1");
             $cmdLogin->bindParam(1, $usuario, PDO::PARAM_STR);
             $cmdLogin->bindParam(2, $clave, PDO::PARAM_STR);
             $cmdLogin->execute();
